archives/fichiers_inutilises/Fichiers inutilisés/requirements.txt
PyQt5==5.15.10
mysql-connector-python==8.4.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
python-docx==1.1.2
python-pptx==0.6.23
odfpy==1.4.1
pypandoc==1.13
